#54: Error a l'intentar modificar entitats.
Task-URL:https://github.com/GovernIB/pinbal/issues/54

S'ha col·locat un if que controla que quant es va a borrar l'organisme
cessionari no sigui null.

+++ pinbal-scsp/src/main/java/es/caib/pinbal/scsp/ScspHelper.java
@@ -546,9 +546,11 @@ public class ScspHelper {
 		OrganismoCesionario organismoCesionario = getOrganismoCesionarioAmbCif(cif);
 		List<ServicioOrganismoCesionario> servicios = getServicioOrganismoCesionarioDao().selectHistorico(organismoCesionario);
 		for (ServicioOrganismoCesionario servicio: servicios) {
-			getPinbalDao().delete(servicio);
+			getPinbalDao().delete(servicio);	
 		}
-		getPinbalDao().delete(organismoCesionario);
+		if(organismoCesionario != null) {
+			getPinbalDao().delete(organismoCesionario);	
+		}
 	}
 	public OrganismoCesionario organismoCesionarioFindByCif(String cif) {
 		LOGGER.debug("Consulta organismo cesionario (" +
